Added comments and modularized some code
