+++ ursa/graph/local_edges.py
@@ -1,7 +1,7 @@
 import ray
 import numpy as np
 from .utils import _filter_remote
-import collections
+from .utils import flatten
 
 MAX_SUBLIST_SIZE = 10
 SAMPLING_COEFFICIENT = 0.1
@@ -46,8 +46,6 @@ class LocalEdges(object):
         @return: A list of edges including the new appended values.
         """
         temp_edges = self.edges[:]
-        # TODO: might need to change the below assigments to how it is done in
-        # the __init__ function
         temp_edges.extend([val for val in values
                            if type(val) is ray.local_scheduler.ObjectID])
         temp_buf = \
@@ -79,6 +77,12 @@ class LocalEdges(object):
         return new_edges
 
     def prune_buffer(self, input_buffer):
+        """Remove duplicates from the buffer of local edges.
+
+        @param input_buffer: The buffer of local edges to be cleaned.
+
+        @return: A list of unique local edges from the buffer.
+        """
         temp_set = set()
         temp_buf = np.array([])
 
@@ -92,20 +96,39 @@ class LocalEdges(object):
         return temp_buf
 
     def get_global_sampling(self):
+        """Get a sorted subsample of edges across all local edges
+
+        @return: A sorted random sample of local edges.
+        """
         num_samples = int(SAMPLING_COEFFICIENT * MAX_SUBLIST_SIZE)
         sample_list = [np.random.choice(ray.get(list_oid), num_samples)
                        for list_oid in self.edges]
-        sample_list = self.flatten(sample_list)
+        sample_list = flatten(sample_list)
         sample_list.sort()
         return sample_list
 
     def get_partitions(self):
+        """Determine the global bounds for partitioning the local edges.
+
+        @return: A list of partition bounds on which to separate the local
+                 edges.
+        """
         indices = [int(i*(SAMPLING_COEFFICIENT * MAX_SUBLIST_SIZE))
                    for i in range(len(self.edges))]
         sample_list = self.get_global_sampling()
         return [sample_list[i] for i in indices]
 
     def partition_sublists(self, list_oid, partition_bounds):
+        """Split all local edges across all sublists in the object store.
+
+        @param list_oid: An OID which points to a list of local edges in the
+                         Ray object store.
+        @param partition_bounds: The list of bounds on which to partition each
+                                 local edge sublist
+
+        @return: A list of sublists, each of which correspond to a partition of
+                 all local edges.
+        """
         partitioned_sublist = [[] for i in range(len(partition_bounds) + 1)]
         for edge in ray.get(list_oid):
             if edge <= partition_bounds[0]:
@@ -119,8 +142,12 @@ class LocalEdges(object):
         return partitioned_sublist
 
     def merge_common_partitions(self):
-        # TODO: How to send this function to a remote object - is it just pass
-        # oid as argument?
+        """Merge common partitions from separate local edge sublists, then sort
+           each merged sublist and push to ray object store
+
+        @return: A list of OID's and buffered edges where all edges in the
+                 object store are globally sorted.
+        """
         if len(self.edges) == 0:
             self.buf.sort()
             return self.buf
@@ -136,16 +163,7 @@ class LocalEdges(object):
             for j in range(len(merged_oid_groupings)):
                 new_partition.extend(merged_oid_groupings[j][i])
             new_partition.sort()
-            new_partition_oid = ray.put(new_partition)
-            new_local_edges.append(new_partition_oid)
+            new_partition = self.prune_buffer(new_partition)
+            new_local_edges.append(ray.put(new_partition))
         new_local_edges.extend(self.buf)
         return new_local_edges
-
-    def flatten(self, x):
-        result = []
-        for el in x:
-            if isinstance(x, collections.Iterable) and not isinstance(el, str):
-                result.extend(self.flatten(el))
-            else:
-                result.append(el)
-        return result
