+++ ursa/graph/utils.py
@@ -3,6 +3,7 @@ from ray import pyarrow as pa
 import numpy as np
 import os
 import ursa
+import collections
 
 
 @ray.remote
@@ -120,3 +121,13 @@ def read_vertex(file):
                                            data["transaction_id"])
 
     return new_vertex
+
+
+def flatten(x):
+    result = []
+    for el in x:
+        if isinstance(x, collections.Iterable) and not isinstance(el, str):
+            result.extend(flatten(el))
+        else:
+            result.append(el)
+    return result
